# Use dot product for containment