--- a/flox/core.py
+++ b/flox/core.py
@@ -302,6 +302,32 @@ def find_group_cohorts(
         for lab in range(bitmask.shape[-1])
     }
 
+    # These invert the label_chunks mapping so we know which labels occur together.
+    def invert(x) -> tuple[np.ndarray, ...]:
+        arr = label_chunks[x]
+        return tuple(arr)
+
+    # No merging is possible when
+    # 1. Our dataset has chunksize one along the axis,
+    single_chunks = all(all(a == 1 for a in ac) for ac in chunks)
+    # 2. Every chunk only has a single group, but that group might extend across multiple chunks
+    one_group_per_chunk = (bitmask.sum(axis=LABEL_AXIS) == 1).all()
+    # 3. Every group is contained to one block, we should be using blockwise here.
+    every_group_one_block = (chunks_per_label == 1).all()
+    # 4. Existing cohorts don't overlap, great for time grouping with perfect chunking
+    # no_overlapping_cohorts = (np.bincount(np.concatenate(tuple(chunks_cohorts.keys()))) == 1).all()
+
+    if (
+        every_group_one_block
+        or one_group_per_chunk
+        or single_chunks
+        # or no_overlapping_cohorts
+        or not merge
+    ):
+        # return chunks_cohorts
+        chunks_cohorts = tlz.groupby(invert, label_chunks.keys())
+        return chunks_cohorts
+
     asfloat = bitmask.astype(float).tocsr()
     containment = (asfloat.T @ asfloat) / chunks_per_label[present_labels]